Simplify Edge & File API
