--- analyzer/restapi-plugin/src/test/java/eu/fasten/analyzer/restapiplugin/mvn/api/impl/FileApiServiceImplTest.java
@@ -1,67 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package eu.fasten.analyzer.restapiplugin.mvn.api.impl;
-
-import eu.fasten.analyzer.restapiplugin.mvn.KnowledgeBaseConnector;
-import eu.fasten.analyzer.restapiplugin.mvn.RestApplication;
-import eu.fasten.core.data.metadatadb.MetadataDao;
-import eu.fasten.core.maven.data.PackageVersionNotFoundException;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class FileApiServiceImplTest {
-
-    private FileApiServiceImpl service;
-    private MetadataDao kbDao;
-    private final int offset = 0;
-    private final int limit = Integer.parseInt(RestApplication.DEFAULT_PAGE_SIZE);
-
-    @BeforeEach
-    void setUp() {
-        service = new FileApiServiceImpl();
-        kbDao = Mockito.mock(MetadataDao.class);
-        KnowledgeBaseConnector.kbDao = kbDao;
-    }
-
-    @Test
-    void getPackageFilesPositiveTest() {
-        var packageName = "pkg";
-        var version = "pkg ver";
-        var response = "files";
-        Mockito.when(kbDao.getPackageFiles(packageName, version, offset, limit)).thenReturn(response);
-        var expected = new ResponseEntity<>(response, HttpStatus.OK);
-        var result = service.getPackageFiles(packageName, version, offset, limit, null, null);
-        assertEquals(expected, result);
-        Mockito.verify(kbDao).getPackageFiles(packageName, version, offset, limit);
-    }
-
-    @Test
-    void getPackageFilesIngestionTest(){
-        var packageName = "junit:junit";
-        var version = "4.12";
-        Mockito.when(kbDao.getPackageFiles(packageName, version, offset, limit)).thenThrow(new PackageVersionNotFoundException("Error"));
-        var result = service.getPackageFiles(packageName, version, offset, limit, null, null);
-        assertEquals(HttpStatus.CREATED, result.getStatusCode());
-        Mockito.verify(kbDao).getPackageFiles(packageName, version, offset, limit);
-    }
-}
